Change asciidoc convention to write a file
This will make this version incompatible with previous ones
for all asciidoc documents.

Previously, a caption of the form:

.Hello: world!
[source,$ext]

would output the code to the file Hello.$ext. Now one need to
write:

.Hello.$ext: world!
[source,$ext]

(where $ext must correspond to the language declared as source)
to obtain the same result.

This is to allow definitions of expansion points to be followed by
a colon and comment as well.